Move deploy API outside of metrics
