+++ lib/sidekiq/api.rb
@@ -6,7 +6,6 @@ require "zlib"
 require "set"
 require "base64"
 
-require "sidekiq/metrics/deploy"
 require "sidekiq/metrics/query"
 
 #
